move torch to separate folder, not required
